project structure modified : adding a jar module so to separate batch(jar) from rest controller(war).
