+++ idsteptodeletetest/pom.xml
@@ -9,29 +9,19 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>fr.abes</groupId>
-    <artifactId>testrestcontrollerinterface</artifactId>
+    <artifactId>idsteptodeletetest</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>testrestcontrollerinterface</name>
+    <name>idsteptodeletetest</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
         <java.version>1.8</java.version>
     </properties>
-    <packaging>war</packaging>
 
     <dependencies>
-
-
-        <!--<dependency>
-            <groupId>fr.abes</groupId>
-            <artifactId>idsteptodeletetest</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>-->
-
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
+            <artifactId>spring-boot-starter</artifactId>
         </dependency>
 
         <dependency>
@@ -46,26 +36,9 @@
         </dependency>
 
         <dependency>
-            <groupId>oracle.ojdbc</groupId>
-            <artifactId>ojdbc</artifactId>
-            <version>12.1.0.1.0</version>
-        </dependency>
-
-        <dependency>
-            <groupId>oracle.ojdbc</groupId>
-            <artifactId>xdb6</artifactId>
-            <version>12.1.0.1.0</version>
-        </dependency>
-
-        <dependency>
-            <groupId>oracle.ojdbc</groupId>
-            <artifactId>xmlparserv2</artifactId>
-            <version>11.2.0.3.0</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-batch</artifactId>
+            <groupId>javax.inject</groupId>
+            <artifactId>javax.inject</artifactId>
+            <version>1</version>
         </dependency>
 
         <dependency>
@@ -80,38 +53,18 @@
             </exclusions>
         </dependency>
         <dependency>
-            <groupId>javax.inject</groupId>
-            <artifactId>javax.inject</artifactId>
-            <version>1</version>
-            <scope>compile</scope>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-batch</artifactId>
         </dependency>
-        <dependency>
+
+
+        <!--<dependency>
             <groupId>fr.abes</groupId>
-            <artifactId>idsteptodeletetest</artifactId>
+            <artifactId>testrestcontrollerinterface</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
-        </dependency>
+        </dependency>-->
     </dependencies>
-
-    <repositories>
-        <repository>
-            <id>artifactory-public</id>
-            <name>artifactory-public</name>
-            <url>https://artifactory.abes.fr/artifactory/maven-public/</url>
-        </repository>
-        <repository>
-            <id>artifactory-public-ext</id>
-            <name>artifactory-public-ext</name>
-            <url>https://artifactory.abes.fr/artifactory/ext-release-local/</url>
-        </repository>
-    </repositories>
-    <pluginRepositories>
-        <pluginRepository>
-            <id>artifactory-public</id>
-            <name>artifactory-public</name>
-            <url>https://artifactory.abes.fr/artifactory/maven-public/</url>
-        </pluginRepository>
-    </pluginRepositories>
 
     <build>
         <plugins>
