+++ idsteptodeletetest/src/main/java/fr/abes/idsteptodeletetest/MessageUsage.java
@@ -1,0 +1,4 @@
+package fr.abes.idsteptodeletetest;
+
+public class MessageUsage {
+}
