+++ idsteptodeletetest/src/main/java/fr/abes/idsteptodeletetest/config/CustomOracle12cDialect.java
@@ -1,0 +1,14 @@
+package fr.abes.idsteptodeletetest.config;
+
+import org.hibernate.dialect.Oracle12cDialect;
+
+
+public class CustomOracle12cDialect extends Oracle12cDialect {
+
+    @Override
+    public boolean useInputStreamToInsertBlob() {
+        //This forces the use of CLOB binding when inserting
+        return false;
+    }
+}
+
