+++ idsteptodeletetest/src/main/java/fr/abes/idsteptodeletetest/star/repositories/StarRepository.java
@@ -1,0 +1,16 @@
+package fr.abes.idsteptodeletetest.star.repositories;
+
+import fr.abes.idsteptodeletetest.star.entities.DocumentStar;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StarRepository extends JpaRepository<DocumentStar, Integer> {
+
+
+    @Query("select s.doc from DocumentStar s where s.iddoc like :x")
+    public String getTefByIddoc(@Param("x") int iddoc);
+
+}
