--- src/main/java/fr/abes/idsteptodelete/MessageUsage.java
@@ -1,4 +1,0 @@
-package fr.abes.idsteptodelete;
-
-public class MessageUsage {
-}
