--- src/main/java/fr/abes/idsteptodelete/ServletInitializer.java
@@ -1,13 +1,0 @@
-package fr.abes.idsteptodelete;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-        return application.sources(IdsteptodeleteApplication.class);
-    }
-
-}
