--- src/main/java/fr/abes/idsteptodelete/config/CustomOracle12cDialect.java
@@ -1,14 +1,0 @@
-package fr.abes.idsteptodelete.config;
-
-import org.hibernate.dialect.Oracle12cDialect;
-
-
-public class CustomOracle12cDialect extends Oracle12cDialect {
-
-    @Override
-    public boolean useInputStreamToInsertBlob() {
-        //This forces the use of CLOB binding when inserting
-        return false;
-    }
-}
-
