+++ src/main/resources/data.csv
@@ -1,0 +1,1 @@
+100010,
