--- src/main/resources/idsteptodelete.txt
@@ -1,3 +1,0 @@
-100009,
-100011,
-100012,
